+++ addin/PyxelRestAddIn/Configuration.cs
@@ -106,12 +106,17 @@ namespace PyxelRestAddIn
             return services;
         }
 
-        public void Save(string configurationFilePath)
+        /// <summary>
+        /// Save services to disk and return a list of python modules that needs to be installed for the configuration to work fine.
+        /// </summary>
+        /// <param name="configurationFilePath"></param>
+        /// <returns>Set of python modules to be installed.</returns>
+        public ISet<string> Save(string configurationFilePath)
         {
             if (configurationFilePath == null)
             {
                 Log.Error("Configuration cannot be saved as configuration file path was not provided.");
-                return;
+                return new HashSet<string>();
             }
             if (!File.Exists(configurationFilePath))
             {
@@ -123,14 +128,19 @@ namespace PyxelRestAddIn
             var parser = new YamlStream();
             var mapping = new YamlMappingNode();
 
+            var pythonModules = new List<string>();
             foreach (Service service in services)
+            {
                 mapping.Add(new YamlScalarNode(service.Name), service.ToConfig());
+                pythonModules.AddRange(service.GetPythonModules());
+            }
 
             parser.Add(new YamlDocument(mapping));
             var writer = new StreamWriter(configurationFilePath);
             parser.Save(writer);
             writer.Close();
             Log.Info("Services configuration updated.");
+            return new HashSet<string>(pythonModules);
         }
 
         internal Service AddDefaultService(string name)